Refactor: Centralize tour logic into tour-core.js
Extract monolithic inline tour DOM manipulation, state management,
and styling out of slashCommands.js into a reusable Tour class.
